:rewind: revert refactor `testing` to `plugins`
pinned for later
