+++ examples/app/zero_shot_agent.py
@@ -5,7 +5,7 @@ from langchain.agents import AgentExecutor, AgentType, initialize_agent, load_to
 from langchain.chat_models import ChatOpenAI
 
 from lanarky import LangchainRouter
-from lanarky.plugins import mount_gradio_app
+from lanarky.testing import mount_gradio_app
 
 load_dotenv()
